Improve test coverage of PolicyFinder
The specs here were "Stubbing the system under test", which is generally
considered to not be good practice, since it can result in 'Imaginary'
specs which don't exercise possible behaviours:

  https://thoughtbot.com/blog/don-t-stub-the-system-under-test

In this case it's not actually possible for the private `#find` method
to return nil.

This PR removes that stubbing and completes a set of test cases which
execute all the paths.

The case where we get a nil object should in normal cases result in an
error, but for some reason in this test suite we have a NilClassPolicy.
This spec was kept in order to be consistent with the corresponding
scope spec.

Other fixes:

- In the test data it seems like `policy_class` and `model_name` should
  only be defined on classes, not instances
- Added test cases to cover https://github.com/varvet/pundit/pull/475